refactor(phase-48.1): split solute_writer.py — gromacs_writer is now a thin shim
- Move write_solute_gro_file + write_solute_top_file to solute_writer.py (780 lines, RISKIEST)
- Empty-molecule_index fallback (lines 3481-3498) preserved verbatim
- 5 MoleculeIndex synthetic entry sites preserved (Group 8 fix — ice/water/guest/custom/solute)
- EXISTING try/except cleanup preserved verbatim (48.1-06 finding: writer HAS it, contrary to research §3 — do NOT add, do NOT remove)
- custom_active 4-writer gate preserved (NOT unified with interface's if custom_guest_info: NOR multi_molecule's custom_by_moltype dict)
- DeprecationWarning for legacy dict preserved in BOTH writers
- _write_top_defaults Format C preserved (comb-rule=2 in helper)
- _write_gro_header + _write_gro_box_vectors + _wrap_aux_positions called (solute uses standard :5d + 9-value triclinic + aux wrapping — NOT divergent)
- gromacs_writer.py is now a THIN re-export shim (787 -> 87 lines, net -700)
- ALL 12 writers moved to per-structure modules (Waves 3-8 COMPLETE)
- 14 byte-equivalence + 18 gmx grompp + 7 solute grompp + 50 scan-all-files + 112 tests/test_output + 75 e2e solute/ion/custom/pbc regression ALL PASS (1840 total tests, no loss)
- ZERO test files updated — Wave 1 + Wave 6 scan-all-files patterns automatically cover the new solute_writer.py
